% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b64Map.R
\name{b64Map}
\alias{b64Map}
\title{b64Map}
\usage{
b64Map(x)
}
\arguments{
\item{x}{(character or numeric) A vector of encoding characters, bit-patterns
or indices.}
}
\value{
(character) A vector of Base 64 characters or their corresponding
  bit-patterns.
}
\description{
\code{b64Map} maps 6-character bit patterns to thier corresponding
Base 64 characters or characters back to bit patterns.
}
\section{Description}{
 if the input is a vector of 6-character bit patterns,
  \code{b64Map} returns the corresponding character from the Base 64
  binary-to-text encoding. If the input is a vector of 1-character strings
  \code{b64Map} returns the corresponding bit pattern. If the input is
  numeric, it is interpreted as indices into the b64 vector. The returned
  vector has the same length as the input but no checking for valid input is
  done.
}

\section{Base 64}{
 Base 64 is a binary to text encoding specified in \href{https://tools.ietf.org/html/rfc4648}{RFC 1738}; we use the URL and filename-safe alphabet version with the following encoding from \code{000000} to \code{111111}:
  \code{"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_"}
}

\examples{
# b64Map three bit patterns
b64Map(c("101101", "011011", "010110"))   # "t"    "b"    "W"

# b64Map three encoding characters
b64Map(c("a", "b", "c"))  # "011010" "011011" "011100"

# print the entire r64 vector as one string
paste0(b64Map(1:64), collapse = "")

}
\author{
\href{https://orcid.org/0000-0002-1134-6758}{Boris Steipe} (aut)
}
