% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq2uu.R
\name{qq2uu}
\alias{qq2uu}
\title{qq2uu.}
\usage{
qq2uu(qq)
}
\arguments{
\item{qq}{(character) a vector of QQIDs}
}
\value{
(character)  a vector of UUIDs
}
\description{
\code{qq2uu} converts a vector of QQIDs to UUIDs.
}
\section{QQIDs}{
 QQIDs are specially formatted UUIDs. See \code{\link{uu2qq}}
  for the the motivation of mapping UUIDs to QQIDs and details on how QQIDs
  are structured. This function reverses the mapping exactly to recover the
  original UUID.
}

\section{Process}{
 To convert a QQID to a UUID, the two "Q-words" that head
  the QQID are mapped to their index in the 1024-element Q-word vector (cf.
  \code{\link{qMap}}), and the indices are converted to two ten bit numbers.
  These twenty bits are expressed as a five-digit hexadecimal number which
  replaces the two Q-words to recover the UUID. For details on UUID format
  see \code{\link{isValidUUID}}. The remaining 18 Base 64 encoded characters
  are converted to their corresponding 27 hex digits via an intermediate
  mapping to bit-patterns.
}

\section{Endianness}{
 The package \code{qqid} uses its own internal functions
  for bit manipulations and is not affected by the endianness of the
  processor architecture it runs on.
}

\examples{
# Convert three example QQIDs and one NA to the corresponding UUIDs
qq2uu( c(QQIDexample(c(1, 3, 5)), NA) )

# Check that the example QQIDs are correctly converted
qq2uu( QQIDexample(2:4) ) == UUIDexample(2:4)  # TRUE TRUE TRUE

}
\seealso{
\code{\link{uu2qq}} to convert a vector of UUIDs to QQIDs.
}
\author{
\href{https://orcid.org/0000-0002-1134-6758}{Boris Steipe} (aut)
}
