% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqMap.R
\name{qqMap}
\alias{qqMap}
\title{qqMap}
\usage{
qqMap(x)
}
\arguments{
\item{x}{(character or numeric) A vector.}
}
\value{
(numeric or character) Indices or QQwords.
}
\description{
\code{<function>} maps numbers to QQ words, or words to their index
                  in the QQ table.
}
\details{
Details.
}
\section{Description}{
 Map words to their index in the QQ table, or indices
                      to the corresponding  word. If x is of type character,
                      the corresponding indices are returned. If x is of
                      type numeric, the corresponding words are returned.
                      The returned vector has the same length as x. Numbers
                      that are not in [1, 2, ... 1024] are mapped to NA.
                      Strings that are not a QQword are mapped to NA.

                      More details on the QQ words.
}

\examples{
# qqMap a number
qqMap(664)   # "joke"
# qqMap three words
qqMap(c("three", "free", "cold", "beer"))  # NA 478  23 261
# return the entire QQ table
x <- qqMap(1:1024)

}
\author{
\href{https://orcid.org/0000-0002-1134-6758}{Boris Steipe} (aut)
}
