% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uu2qq.R
\name{uu2qq}
\alias{uu2qq}
\title{uu2qq.}
\usage{
uu2qq(uu)
}
\arguments{
\item{uu}{(character) ...}
}
\value{
(character)  ...
}
\description{
\code{uu2qq} convert a vector of UUIDs to QQIDs.
}
\details{
Details.
}
\section{<title>}{
 Additional explanation.
}

\examples{
# <explain what the example does>
uu2qq(UUIDexample()[c(1, 3, 5)])

}
\seealso{
\code{\link{qq2uu}} to convert a vector of QQIDs to UUIDs.
}
\author{
\href{https://orcid.org/0000-0002-1134-6758}{Boris Steipe} (aut)
}
