% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UUIDexample.R
\name{UUIDexample}
\alias{UUIDexample}
\title{UUIDexample}
\usage{
UUIDexample()
}
\value{
(character) Five UUIDs
}
\description{
\code{UUIDexample} returns five UUIDs.
}
\details{
Details.
The same, fixed five UUIDs are returned every time. They can be mapped
to the five QQIDs returned by \code{QQIDexample}.
}
\examples{
UUIDexample()

}
\seealso{
\code{\link{QQIDexample}} Returns five QQIDs
}
\author{
\href{https://orcid.org/0000-0002-1134-6758}{Boris Steipe} (aut)
}
