% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidQQID.R
\name{isValidQQID}
\alias{isValidQQID}
\title{isValidQQID}
\usage{
isValidQQID(s, na.map = FALSE)
}
\arguments{
\item{s}{(character) a vector of strings to check.}

\item{na.map}{(logical) replace NA with \code{FALSE} (default), \code{NA},
or  \code{TRUE}}
}
\value{
(logical) a vector of the same length as the input, \code{TRUE}
                  for every element  of the input that is a valid QQID.
}
\description{
\code{isValidQQID} Test whether the function argument is a vector of valid
                   QQIDs.
}
\details{
The function accepts a vector of strings and returns a logical
vector of the same length,  \code{TRUE} for every element of the input that
is a valid QQID. \code{NA} values are mapped to \code{FALSE} (default) or can
be replaced with \code{NA} to preserve them. Note: arguments passed to
\code{na.map} are implicitly converted to type logical.
}
\examples{
# check 2 valid QQIDs
isValidQQID(QQIDexample()[1:2])   # TRUE  TRUE
# check one invalid QQID
isValidQQID("no.pasaran-df7-652d-1982-0a5c-2c1a88284ee8")

}
\seealso{
\code{\link{isValidUUID}} to check UUIDs.
}
\author{
\href{https://orcid.org/0000-0002-1134-6758}{Boris Steipe} (aut)
}
