% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qq2uu.R
\name{qq2uu}
\alias{qq2uu}
\title{qq2uu.}
\usage{
qq2uu(qq)
}
\arguments{
\item{qq}{(character) ...}
}
\value{
(character)  ...
}
\description{
\code{qq2uu} convert a vector of QQIDs to UUIDs.
}
\details{
Details.
}
\section{<title>}{
 Additional explanation.
}

\examples{
# <explain what the example does>
qq2uu(QQIDexample()[c(1, 3, 5)])

}
\seealso{
\code{\link{uu2qq}} to convert a vector of UUIDs to QQIDs.
}
\author{
\href{https://orcid.org/0000-0002-1134-6758}{Boris Steipe} (aut)
}
