% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QQIDexample.R
\name{QQIDexample}
\alias{QQIDexample}
\title{QQIDexample}
\usage{
QQIDexample()
}
\value{
(character) Five QQIDs
}
\description{
\code{QQIDexample} returns five QQIDs.
}
\details{
Details.
The same, fixed five QQIDs are returned every time. They can be mapped
to the five UUIDs returned by \code{UUIDexample}.
}
\examples{
QQIDexample()

}
\seealso{
\code{\link{UUIDexample}} Returns five UUIDs
}
\author{
\href{https://orcid.org/0000-0002-1134-6758}{Boris Steipe} (aut)
}
