% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UUIDexample.R
\name{UUIDexample}
\alias{UUIDexample}
\title{UUIDexample}
\usage{
UUIDexample(idx = 1:5)
}
\arguments{
\item{idx}{(numeric) an index vector that defines which of the five
internally stored example UUIDs to return}
}
\value{
(character) a vector of UUIDs
}
\description{
\code{UUIDexample} returns synthetic, valid UUIDs for testing and
development, which are easy to distinguish from "real" UUIDs to prevent their
accidental use as IDs.
}
\details{
The function stores five artificial sample UUIDs. Input is an index vector
that specifies which UUIDs to return. More than five UUIDs can be requested
by repeating indices. The UUIDs can be converted to the QQIDs provided by
\code{QQIDexample}. Note: these are not random numbers in the sense of the
philosophy behind UUIDs. Do not use them for any purpose other than
demonstration and testing.
}
\examples{
UUIDexample()                             # the five stored UUIDs
UUIDexample(2:3)                          # two UUIDs
UUIDexample(rep(1:5, 5))                  # twentyfive non-random UUIDS
UUIDexample(3) == qq2uu(QQIDexample(3))   # TRUE (testing the conversion)

}
\seealso{
\code{\link{QQIDexample}} Returns five QQIDs
}
\author{
\href{https://orcid.org/0000-0002-1134-6758}{Boris Steipe} (aut)
}
